+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-### Unreleased
+### [v0.225](https://github.com/arez/arez/tree/v0.225) (2026-01-23) · [Full Changelog](https://github.com/arez/arez/compare/v0.224...v0.225) · [API Differences](https://arez.github.io/api-diff?key=arez&old=0.224&new=0.225)
+
+The release includes 1 non breaking API change.
+
+Changes in this release:
 
 * Reject `@Action(reportParameters = false)` when the action method has no parameters.
 * Add `@ObservableInitial` to provide initial values for abstract observables.